requirements.txt
aes==1.0.0
allure-pytest==2.8.6
allure-python-commons==2.8.6
amqp==1.4.9
anyjson==0.3.3
apipkg==1.5
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
crypto==1.4.1
cryptography==2.8
Django==2.0.3
django-celery==3.2.2
eventlet==0.22.1
execnet==1.7.1
flower==0.9.2
get==2019.4.13
greenlet==0.4.15
har2case==0.3.1
httpie==1.0.3
HttpRunner==1.5.8
idna==2.6
importlib-metadata==1.4.0
Jinja2==2.11.1
kombu==3.0.37
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.1.0
Naked==0.1.31
packaging==20.0
paramiko==2.4.1
pluggy==0.13.1
post==2019.4.13
public==2019.4.13
py==1.8.1
pyasn1==0.4.8
pycparser==2.19
pycryptodome==3.9.6
Pygments==2.5.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.6
pytest==5.3.5
pytest-forked==1.1.3
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-rerunfailures==8.0
pytest-xdist==1.23.2
pytz==2019.3
PyYAML==3.12
query-string==2019.4.13
request==2019.4.13
requests==2.18.4
requests-toolbelt==0.9.1
selenium==3.141.0
shellescape==3.8.1
six==1.14.0
tornado==6.0.3
urllib3==1.22
wcwidth==0.1.8
zipp==1.0.0
